+++ app/serializers/dc20/character_serializer.rb
@@ -7,14 +7,14 @@ module Dc20
                :skills, :skill_points, :skill_expertise_points, :guide_step, :trade_points, :trade_expertise_points,
                :language_points, :trades, :trade_knowledge, :language_levels, :attribute_saves, :physical_save, :mental_save,
                :initiative, :conditions, :attacks, :health, :path_points, :paths, :mana_points, :maneuvers, :maneuver_points,
-               :stamina_points, :grit_points, :rest_points, :mana_spend_limit, :cantrips, :spells
+               :stamina_points, :grit_points, :rest_points, :mana_spend_limit, :cantrips, :spells, :talent_points, :talents
 
     delegate :features, :id, :name, :level, :main_class, :abilities, :modified_abilities, :health, :attribute_points, :classes,
              :ancestries, :combat_mastery, :save_dc, :precision_defense, :area_defense, :attack, :skills, :skill_points,
              :skill_expertise_points, :trade_points, :trade_expertise_points, :language_points, :trades, :trade_knowledge,
              :language_levels, :attribute_saves, :physical_save, :mental_save, :initiative, :conditions, :attacks, :path_points,
              :paths, :mana_points, :maneuvers, :maneuver_points, :stamina_points, :grit_points, :rest_points,
-             :mana_spend_limit, :cantrips, :spells, to: :decorator
+             :mana_spend_limit, :cantrips, :spells, :talent_points, :talents, to: :decorator
     delegate :created_at, :updated_at, :data, to: :object
     delegate :guide_step, to: :data